change project

+++ src/main/java/com/example/api/generic/GenericController.java
@@ -4,6 +4,7 @@ import java.io.Serializable;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.EmptyResultDataAccessException;
 
 public class GenericController<T, ID extends Serializable> {
 
@@ -27,7 +28,17 @@ public class GenericController<T, ID extends Serializable> {
 		return repository.saveAndFlush(t);
 	}
 	
-	protected void deleteByid(Long id) {
-		repository.delete(id);
+	protected void deleteByid(Long id) throws EmptyResultDataAccessException {
+		try {
+			repository.delete(id);
+		} catch (EmptyResultDataAccessException ex) {
+			throw new EmptyResultDataAccessException(ex.getMessage(), Integer.valueOf(id.toString()));
+		}
+	}
+	
+	protected void deleteByGroup(List<ID> ids) {
+		ids.forEach(id -> {
+			if(repository.exists(id)) repository.delete(id);
+		});
 	}
 }
